feat: add proper SolidActor and OpenHabActor
